+++ android-app/app/src/test/java/com/lorenzo/mangadownloader/SettingsStoreTest.kt
@@ -39,7 +39,6 @@ class SettingsStoreTest {
             autoDownloadBatchSize = 2,
             smartCleanupEnabled = true,
             smartCleanupKeepPreviousChapters = 1,
-            streamingReaderEnabled = true,
             parentalControlEnabled = true,
             parentalPinConfigured = true,
             parentalBiometricEnabled = true,